Workflow parsers

See merge request nomad-lab/nomad-FAIR!379

+++ nomad/datamodel/metainfo/simulation/calculation.py
@@ -1472,7 +1472,7 @@ class BaseCalculation(MSection):
         categories=[FastAccess])
 
     n_references = Quantity(
-        type=np.dtype(np.float64),
+        type=np.dtype(np.int32),
         shape=[],
         description='''
          Number of references to the current section calculation.
